refactor: 서버에서 생성하고 저장되는 id값을 통해 폴더 요청하도록 변경

+++ server/src/main/java/wap/starlist/bookmark/controller/BookmarkController.java
@@ -107,7 +107,7 @@ public class BookmarkController {
     }
 
     @GetMapping("/nested/{id}")
-    public ResponseEntity<?> getNestedNodes(@AuthenticationPrincipal String loginUser, @PathVariable("id") Integer id) {
+    public ResponseEntity<?> getNestedNodes(@AuthenticationPrincipal String loginUser, @PathVariable("id") Long id) {
         List<BookmarkNodeResponse> nodes = bookmarkService.getNestedNodes(loginUser, id);
         return ResponseEntity.ok().body(nodes);
     }
